Modified subnet_cadmium.hpp acc. to coding conventions
